Refactor entry point to support dependency injection (#45)
This commit defines a new abstract class, DynSemEntryPoint, which is
responsible for instantiating the VM with the proper dependencies: a
parser, specification term, and a term registry.

The dependencies are injected by passing them as "configuration" to the
VM builder (see PolyglotEngine.Builder#config). The configuration
parameters are then retrieved when creating the DynSemContext in
DynSemLanguage#createContext.

This allows to inject ones own "parsers", for example a parser which
does not parse, but just returns an ATerm. Thus an entry point can be
created which evaluates ATerms instead of just program source code (see
issue #43).