Add injection of a RuleRegistry in the entry point
The dynsem.properties file allows for optionally specifying a custom
RuleRegistry, so to support this it needs to be injected when creating
the VM.

To populate the RuleRegistry, the DynSemContext.LANGUAGE singleton needs
to be set. Thus an assertion has been added to
DynSemEntryPoint#getCallable, and the implementing class (i.e. the
generated entry point) needs to set this as a precondition.