Lookup resources in the classpath instead of current working directory
This way, any project that includes the generated interpreter as a .jar
can reuse these methods given that the .jar archive is on the classpath

This comes with one assumption, namely that the specification term file
and the parse table are put into a directory which are copied to the
.jar archive or elsewhere on the classpath upon compilation.

For simpl this is already the case: e.g. it specifies
"src/main/resources/specification.aterm" as "target.package". Maven then
makes sure that files in "src/main/resources" are put into
"target/classes" and the .jar archive when building the project.