Refactor to "evaluate" dynsem specification instead of program source
Long commit message incoming, see TL;DR at the bottom for short message.

Since this is a meta-interpreter, it makes sense to pass the
specification term to the eval method, instead of program source for
that interpreter.

We pass the DynSem specification, and use it to fill the rule registry.
The RuleRoots in the rule registry are then available for foreign
access, as is the case with DynSemPrimedRun before this commit. However,
this commit renders DynSemPrimedRun unnecessary, so it has been removed
and replaced with DynSemRule (a wrapper around RuleRoot).

Now a client/library user of the interpreter can invoke any rule by
finding the rule with PolyglotEngine#findGlobalSymbol with a rulekey
(e.g. "init/Program/1"), and then calling the rule with a program term
(ITerm) and (optionally) semantic components.

Since we now parse the specification term instead of program source
code, the parser is no longer needed in the DynSemContext and does not
have to be used by a generated subclass of DynSemLanguage. Parsing is
now completely separate from interpretation and the PolyglotEngine#eval
method, thus improving reusability.

**TL;DR:** In short, the benefits that this commits brings are:

1) The parsing can be done outside of the parse method in the generated
   language class.

2) Because of 1), providing an entry point for ASTs is trivial, since
   the interpreter can now be called directly with ITerms. (issue #43)

3) Any rule can now be invoked by a client that has the generated
   interpreter in the classpath. Previously this was not possible.

4) It makes the whole thing more logical, as it is consistent with the
   rest of the idea of a "meta-interpreter": you pass the interpreter
   specification which results in an interpreter as output.